tests/wpt/tests/css/css-shapes/shape-functions/xywh-function-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the xywh() function</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-xywh">
<meta name="assert" content="Tests parsing of the xywh() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("shape-outside", "xywh(0px 1px 2% 3em)");
test_valid_value("shape-outside", "xywh(0px calc(100% - 20px) 2% 3em)");
test_valid_value("shape-outside", "xywh(10px 20px 30px 25px round 100%)");
test_valid_value("shape-outside", "xywh(10px 20px 30px 25px round 0 1px)", "xywh(10px 20px 30px 25px round 0px 1px)");
test_valid_value("shape-outside", "xywh(10px 20px 30px 25px round 0px 1px 2%)");
test_valid_value("shape-outside", "xywh(10px 20px 30px 25px round 0px 1px 2% 3em)");
test_valid_value("shape-outside", "xywh(10px 20px 30px 25px round 20% / 0px 1px 2% 3em)");
</script>
</body>
</html>
