third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/xywh-function-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['shape-outside'] = "xywh(0px 1px 2% 3em)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(0px calc(100% - 20px) 2% 3em)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(10px 20px 30px 25px round 100%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(10px 20px 30px 25px round 0 1px)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(10px 20px 30px 25px round 0px 1px 2%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(10px 20px 30px 25px round 0px 1px 2% 3em)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "xywh(10px 20px 30px 25px round 20% / 0px 1px 2% 3em)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
